backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pydantic==2.11.2
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-multipart==0.0.20
apscheduler==3.10.4
httpx==0.28.1
reportlab==4.3.1
python-docx==1.1.2
pypdf==6.10.0
neo4j==5.28.1
boto3==1.37.34
pytest==8.3.5
pytest-asyncio==0.26.0
openai==1.75.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
